outputs/requirements.txt
python-telegram-bot==21.6
python-dotenv==1.0.1
openai==1.40.0
pymupdf>=1.23.0
python-docx>=1.1.0
pytesseract>=0.3.10
Pillow>=10.0.0
